STYLE: Fix minor issues in vtkSegmentationCore documentation
- Fix broken links in vtkSegmentationCore documentation
- Fix compiler warnings in vtkSegmentationCore
  Note: the library has been copied to a new repository for later re-use by other projects: https://github.com/PerkLab/PolySeg
- Fix tooltip message in segmentation geometry widget

svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=27295
git-svn-id: http://svn.slicer.org/Slicer4/trunk@27295 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/vtkSegmentationCore/vtkSegmentationConverter.cxx
@@ -310,7 +310,7 @@ vtkSegmentationConverter::ConversionPathType vtkSegmentationConverter::GetCheape
     {
     if (pathIt->second <= cheapestPathCost)
       {
-      unsigned int numberOfConversions = pathIt->first.size();
+      size_t numberOfConversions = pathIt->first.size();
       if (pathIt->second == cheapestPathCost)
         {
         // If the path cost is exactly the same then compare the number of conversions too
@@ -321,7 +321,7 @@ vtkSegmentationConverter::ConversionPathType vtkSegmentationConverter::GetCheape
           }
         }
       cheapestPathCost = pathIt->second;
-      cheapestPathNumberOfConversions = numberOfConversions;
+      cheapestPathNumberOfConversions = (unsigned int)numberOfConversions;
       cheapestPath = pathIt->first;
       }
     }
